Fix: time_entries limit() needs Long, not Int
CI caught it: PostgrestQueryBuilder.limit(count) takes a Long. Existing call sites pass
integer LITERALS (limit(200)) which Kotlin coerces to Long, but getRecent passed an Int
VARIABLE, which does not auto-convert — "actual type Int, but Long was expected". This
single error blocked the whole feature-parity stack (time tracking → QuickBooks → logo →
NEC all build on it); the tip build confirmed it was the ONLY error. Use limit(limit.toLong()).

+++ native-android/app/src/main/java/com/wirewaypro/app/data/timetracking/TimeEntryRepositoryImpl.kt
@@ -38,7 +38,7 @@ class TimeEntryRepositoryImpl @Inject constructor(
             .select {
                 filter { eq("user_id", userId) }
                 order("created_at", Order.DESCENDING)
-                limit(limit)
+                limit(limit.toLong())
             }
             .decodeList<TimeEntryDto>()
             .map { it.toDomain() }
